% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodiff.R
\name{calculate_ed}
\alias{calculate_ed}
\title{Calculate a suite of six emotion differentiation measures}
\usage{
calculate_ed(dat, emotions, ...)
}
\arguments{
\item{dat}{a data.frame object with moment-level observations}

\item{emotions}{A character vector of all relevant emotion variables in dat}

\item{...}{Grouping variable(s). Usually this will just be a person ID. Occasionally, you might have a wave ID and
a person ID, if you want to calculate scores for each individual within each wave.}
}
\value{
the input data.frame with 6 emotion differentiation variables appended
}
\description{
Given a dataframe of emotion data over time, a list of emotion variables in that dataset, and an ID variable,
calculates momentary emotional differentiation, classic emotion differentiation, L2 emotion differentiation, and
their inverses
}
\examples{
calculate_ed(dat = emo_ex, emotions = c("happy","relaxed","cheerful"), id)
}
